Improve 30 game function fuzzy matches

+++ src/game/fsys/fsys_file_r52_8017D410_prefix.c
@@ -1,22 +1,22 @@
 /* Score instrumentation only; not evidence of a retail TU boundary. */
 #include "src/game/fsys/fsys_file_candidate_8017D410.c"
 
-FSYSSlot* fn_8017D410(u32 fileHandle, u32 mode)
+FSYSSlot* fn_8017D410(u32 fileHandle, s32 mode)
 {
     u32 i;
     FSYSSlot* slot;
 
     slot = (FSYSSlot*)lbl_8047B1B4;
     for (i = 0; i < lbl_80453FEC.maxSlots; i++, slot++) {
-        if (slot->status == 0 || slot->fileHandle != fileHandle) {
+        if ((s32)slot->status == 0 || slot->fileHandle != fileHandle) {
             continue;
         }
 
-        if (slot->status != 1000) {
+        if ((s32)slot->status != 1000) {
             if (mode == 2 || mode == 7) {
                 return NULL;
             }
-            if (slot->loadMode == 3) {
+            if ((s32)slot->loadMode == 3) {
                 return slot;
             }
             slot->reloadFlag = 1;
@@ -25,13 +25,13 @@ FSYSSlot* fn_8017D410(u32 fileHandle, u32 mode)
 
         switch (mode) {
         case 0:
-            if (slot->loadMode != 2 && slot->loadMode != 7) {
+            if ((s32)slot->loadMode != 2 && (s32)slot->loadMode != 7) {
                 slot->refCount++;
             }
             break;
         case 2:
         case 7:
-            if (slot->refCount == 0) {
+            if ((s32)slot->refCount == 0) {
                 slot->refCount++;
             }
             break;
@@ -41,7 +41,7 @@ FSYSSlot* fn_8017D410(u32 fileHandle, u32 mode)
 
     slot = (FSYSSlot*)lbl_8047B1B4;
     for (i = 0; i < lbl_80453FEC.maxSlots; i++, slot++) {
-        if (slot->status == 0) {
+        if ((s32)slot->status == 0) {
             slot->refCount = 0;
             slot->refCount++;
             return slot;
